Changing refs to ukprn back to int

+++ src/SFA.DAS.AssessorService.Application.Api.UnitTests/WebAPI/Organisations/Maintenence/Post/Handlers/WhenDeletedOrganisationExists.cs
@@ -24,7 +24,7 @@ namespace SFA.DAS.AssessorService.Application.Api.UnitTests.WebAPI.Organisations
             _orgRepos = new Mock<IOrganisationRepository>();
             
             var orgQueryRepos = new Mock<IOrganisationQueryRepository>();
-            orgQueryRepos.Setup(r => r.GetByUkPrn(It.IsAny<long>())).ReturnsAsync(new OrganisationResponse()
+            orgQueryRepos.Setup(r => r.GetByUkPrn(It.IsAny<int>())).ReturnsAsync(new OrganisationResponse()
             {
                 Status = OrganisationStatus.Deleted,
                 EndPointAssessorOrganisationId = "12345"
